Substituted the old track class with the new one
