+++ Cases/car.json
@@ -1,19 +1,27 @@
 {
-    "case":         "Car",
-    "stepsize":     {
-                    "min": 20,
-                    "max": 30
-                    },
-    "rlim":         200,
-    "fc":           28e9,
-    "receiver":     {
-                    "antennea": 4,
-                    "beams": 5
-                    },
-    "transmitter":  {
-                    "antennea": 4,
-                    "beams": 5
-                    },
-    "P_t":          10000,
-    "change_dir":   0.001
+    "case":             "Car",
+    "rlim":             200,
+    "fc":               28e9,
+    "receiver":         {
+                        "antennea": 4,
+                        "beams": 5
+                        },
+    "transmitter":      {
+                        "antennea": 4,
+                        "beams": 5
+                        },
+    "P_t":              10000,
+    "vmax":             14,
+    "vpref":            [0, 8.3, 14],
+    "pvpref":           [0.3, 0.1, 0.4],
+    "vchange":          50,
+    "dirchange":        50,
+    "stop_dirchange":   0.33,
+    "acc_max":          2.5,
+    "dec_max":          4,
+    "curvetime":        {
+                        "max":  10,
+                        "min":  1
+                        },
+    "static_friction":  0.7
 }
