--- Cases/walk.json
@@ -1,19 +1,0 @@
-{
-    "case":         "Walk",
-    "stepsize":     {
-                    "min": 0.007,
-                    "max": 0.007
-                    },
-    "rlim":         200,
-    "fc":           28e9,
-    "receiver":     {
-                    "antennea": 4,
-                    "beams": 8
-                    },
-    "transmitter":  {
-                    "antennea": 4,
-                    "beams": 8
-                    },
-    "P_t":          10000,
-    "change_dir":   0.0001
-}
