--- New_Walk/car.json
@@ -1,16 +1,0 @@
-{
-    "case":             "Car",
-    "vmax":             14,
-    "vpref":            [0, 8.3, 14],
-    "pvpref":           [0.3, 0.1, 0.4],
-    "vchange":          50,
-    "dirchange":        50,
-    "stop_dirchange":   0.33,
-    "acc_max":          2.5,
-    "dec_max":          4,
-    "curvetime":        {
-                        "max":  10,
-                        "min":  1
-                        },
-    "static_friction":  0.7
-}
